@@ -1,87 +1,95 @@
 import MEGASwift
 @testable import MEGAUIKit
-import XCTest
+import Testing
+import UIKit
 
-final class UIMenuTests: XCTestCase {
-    var action1: UIAction!
-    var action2: UIAction!
+@MainActor
+struct UIMenuTests {
+    let action1 = UIAction(title: "Action 1", handler: {_ in})
+    let action2 = UIAction(title: "Action 2", handler: {_ in})
     
-    override func setUp() {
-        super.setUp()
-        action1 = UIAction(title: "Action 1", handler: {_ in})
-        action2 = UIAction(title: "Action 2", handler: {_ in})
-    }
-    
+    @Test
     func testCompareMenuWhere_EitherAreNil() {
         let menuItemA: UIMenu? = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
         let menuItemB: UIMenu? = nil
-        XCTAssertFalse(menuItemA ~~ menuItemB)
-        XCTAssertFalse(menuItemB ~~ menuItemA)
+        #expect((menuItemA ~~ menuItemB) == false)
+        #expect((menuItemB ~~ menuItemA) == false)
     }
     
+    @Test
     func testCompareMenuWhere_BothHaveSameData() {
         let menuItemA = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
         let menuItemB = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
-        XCTAssertTrue(menuItemA ~~ menuItemB)
+        #expect((menuItemA ~~ menuItemB) == true)
     }
     
+    @Test
     func testCompareMenuWith_DifferentTitle() {
         let menuItemA = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
         let menuItemB = UIMenu(title: "UIMenu.ItemB", image: nil, options: .displayInline, children: [])
-        XCTAssertFalse(menuItemA ~~ menuItemB)
+        #expect((menuItemA ~~ menuItemB) == false)
     }
     
+    @Test
     func testCompareMenuWith_DifferentImage() {
         let menuItemA = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
         let menuItemB = UIMenu(title: "UIMenu.ItemB", image: UIImage(), options: .displayInline, children: [])
-        XCTAssertFalse(menuItemA ~~ menuItemB)
+        #expect((menuItemA ~~ menuItemB) == false)
     }
     
+    @Test
     func testCompareMenuWith_DifferentDisplayOptions() {
         let menuItemA = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
         let menuItemB = UIMenu(title: "UIMenu.ItemB", image: UIImage(), options: .destructive, children: [])
-        XCTAssertFalse(menuItemA ~~ menuItemB)
+        #expect((menuItemA ~~ menuItemB) == false)
     }
     
+    @Test
     func testCompareMenuWith_DifferentChildren() {
         let menuItemA = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [])
         let menuItemB = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [action1])
-        XCTAssertFalse(menuItemA ~~ menuItemB)
+        #expect((menuItemA ~~ menuItemB) == false)
     }
     
+    @Test
     func testCompareMenuWith_SameChildren() {
         let menuItemA = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [action1])
         let menuItemB = UIMenu(title: "UIMenu.ItemA", image: nil, options: .displayInline, children: [action1])
-        XCTAssertTrue(menuItemA ~~ menuItemB)
+        #expect((menuItemA ~~ menuItemB) == true)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusAreIdentical_returnsTrue() {
         let menu = UIMenu(title: "Test Menu", children: [action1, action2])
         
-        XCTAssertTrue(UIMenu.match(lhs: menu, rhs: menu))
+        #expect(UIMenu.match(lhs: menu, rhs: menu) == true)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusAreDifferentButHaveSameActions_returnsTrue() {
         let oldMenu = UIMenu(title: "Old Menu", children: [action1, action2])
         let updatedMenu = UIMenu(title: "Updated Menu", children: [action1, action2])
         
-        XCTAssertTrue(UIMenu.match(lhs: oldMenu, rhs: updatedMenu))
+        #expect(UIMenu.match(lhs: oldMenu, rhs: updatedMenu) == true)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusHaveDifferentActions_returnsFalse() {
         let oldMenu = UIMenu(title: "Old Menu", children: [action1])
         let updatedMenu = UIMenu(title: "Updated Menu", children: [action2])
     
-        XCTAssertFalse(UIMenu.match(lhs: oldMenu, rhs: updatedMenu))
+        #expect(UIMenu.match(lhs: oldMenu, rhs: updatedMenu) == false)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusHaveSameActionsInDifferentOrder_returnsFalse() {
         let oldMenu = UIMenu(title: "Test Menu", children: [action1, action2])
         let updatedMenu = UIMenu(title: "Test Menu", children: [action2, action1])
         
-        XCTAssertFalse(UIMenu.match(lhs: oldMenu, rhs: updatedMenu))
+        #expect(UIMenu.match(lhs: oldMenu, rhs: updatedMenu) == false)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusHaveSameActionsAndActionStates_returnsTrue() {
         let action1_ON = UIAction(title: "Action 1", state: .on, handler: {_ in})
         let action2_OFF = UIAction(title: "Action 2", state: .off, handler: {_ in})
@@ -89,9 +97,10 @@ final class UIMenuTests: XCTestCase {
         let oldMenu = UIMenu(title: "Test Menu", children: [action1_ON, action2_OFF])
         let updatedMenu = UIMenu(title: "Test Menu", children: [action1_ON, action2_OFF])
         
-        XCTAssertTrue(UIMenu.match(lhs: oldMenu, rhs: updatedMenu))
+        #expect(UIMenu.match(lhs: oldMenu, rhs: updatedMenu) == true)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusAreDifferentButHaveSameActionsAndActionStates_returnsFalse() {
         let action1_ON = UIAction(title: "Action 1", state: .on, handler: {_ in})
         let action2_OFF = UIAction(title: "Action 2", state: .off, handler: {_ in})
@@ -99,9 +108,10 @@ final class UIMenuTests: XCTestCase {
         let oldMenu = UIMenu(title: "Old Menu", children: [action1_ON, action2_OFF])
         let updatedMenu = UIMenu(title: "Updated Menu", children: [action1_ON, action2_OFF])
         
-        XCTAssertTrue(UIMenu.match(lhs: oldMenu, rhs: updatedMenu))
+        #expect(UIMenu.match(lhs: oldMenu, rhs: updatedMenu) == true)
     }
     
+    @Test
     func testDoMenuActionMatch_whenMenusAreDifferentAndHaveSameActionsButDifferentActionStates_returnsFalse() {
         let action1_ON = UIAction(title: "Action 1", state: .on, handler: {_ in})
         let action1_OFF = UIAction(title: "Action 1", state: .off, handler: {_ in})
@@ -111,9 +121,10 @@ final class UIMenuTests: XCTestCase {
         let oldMenu = UIMenu(title: "Old Menu", children: [action1_ON, action2_OFF])
         let updatedMenu = UIMenu(title: "Updated Menu", children: [action1_OFF, action2_ON])
         
-        XCTAssertFalse(UIMenu.match(lhs: oldMenu, rhs: updatedMenu))
+        #expect(UIMenu.match(lhs: oldMenu, rhs: updatedMenu) == false)
     }
     
+    @Test
     func testMatch_sameActionStates_returnsFalse() {
         let action1_ON = UIAction(title: "Action 1", state: .on, handler: {_ in})
         let menu1 = UIMenu(title: "Updated Menu", options: [], children: [action1_ON])
@@ -121,9 +132,10 @@ final class UIMenuTests: XCTestCase {
         
         let result = UIMenu.match(lhs: menu1, rhs: menu2)
        
-        XCTAssertTrue(result)
+        #expect(result == true)
    }
     
+    @Test
     func testMatch_differentActionStates_returnsFalse() {
         let action1_ON = UIAction(title: "Action 1", state: .on, handler: {_ in})
         let menu1 = UIMenu(title: "Updated Menu", options: [], children: [action1_ON])
@@ -132,6 +144,6 @@ final class UIMenuTests: XCTestCase {
         
         let result = UIMenu.match(lhs: menu1, rhs: menu2)
        
-        XCTAssertFalse(result)
+        #expect(result == false)
    }
 }
