@@ -1,6 +1,7 @@
 import UIKit
 
-public extension Sequence where Element == Optional<UIView> {
+public extension [UIView?] {
+    @MainActor
     func applySensitiveAlpha(isSensitive: Bool) {
         self.forEach { $0?.alpha = isSensitive ? 0.5 : 1 }
     }
